requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0

# API Clients
google-search-results>=2.4.2

# Data validation and processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Web framework
fastapi>=0.110.0
uvicorn>=0.27.0

# Utilities
numpy>=1.26.0
python-dateutil>=2.8.2
typing-extensions>=4.0.0

# Optional: for better string matching
python-Levenshtein>=0.21.0
